requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
Django==5.1.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
executing==2.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
ipython==8.31.0
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
